Refactor pair and unpair parser names to reflect DG
